Refactor Kotlin Support
