+++ integrations/kotlin-support/nima/src/main/java/module-info.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2023 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * Helidon Nima Kotlin Support
+ */
+module io.helidon.integrations.kotlin.support.nima {
+
+    requires io.helidon.nima.webserver;
+
+    requires java.logging;
+
+    requires kotlin.stdlib;
+
+    exports io.helidon.integrations.kotlin.support.nima;
+}
